pyproject.toml
[tool.poetry]
name = "europythie"
version = "0.1.0"
description = ""
authors = ["Hugo Porchet <hugo.porchet.belmonte@gmail.com>"]
repository = "https://github.com/Ttanesque/EuropPythie"
license = "MIT"
readme = "README.md"
packages = [{include = "euro_pythie"}]

[tool.poetry.dependencies]
python = "^3.11"
interactions-py = "^5.11.0"
black = "^23.12.0"
schedule = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
